fix: the conv leaked bf16 into the long head's triangular solve
Both conv arms crashed on their first forward:

  NotImplementedError: "triangular_solve_cuda" not implemented for 'BFloat16'

My bug, and a clean example of verifying the wrong path. LayerNorm stays fp32 under
autocast, so z has always reached the heads in fp32 and the long head's triangular solve
worked. F.conv1d is autocast-eligible and returns bf16, so inserting the conv changed z's
dtype and handed bf16 to a solve that has no bfloat16 CUDA kernel at all.

The three checks I ran before shipping it -- identity at init, decode == prefill, split
prefill -- all passed, and all three ran in float64 with NO autocast. They proved the
conv's arithmetic and said nothing about the path the training run actually takes. The
fix is one cast back to z.dtype in each of the four conv entry points (prefill and step,
in lapa/layer.py and sca2/arch_short.py).

Re-verified, this time including the path that broke: cshort_damphkv and LaplaceAttention
both run conv=0 and conv=4 under bf16 autocast, forward and backward, output fp32. Plus
the original three, still exact -- conv=4 at init equals conv=0 bit-for-bit (0.00e+00),
decode == prefill 1.1e-15 -- and the standing gates, float64 vs the sca2 mirror 1.8e-16
and sca2.iso cshort_damphkv OK.

d1024_kv_conv4 relaunched and healthy at 19.0k tok/s.

+++ sca2/arch_short.py
@@ -194,7 +194,10 @@ class ShortLayer(SCA2Layer):
         zz = torch.cat([buf.to(z.dtype), z], 1)
         zc = F.conv1d(zz.transpose(1, 2), self.cw.to(z.dtype),
                       groups=self.cfg.d).transpose(1, 2)
-        return zc, zz[:, -(self.ck - 1):]
+        # BACK TO z's dtype. LayerNorm returns fp32 under autocast, but conv1d is an
+        # autocast-eligible op and returns bf16; letting that through hands bf16 to the
+        # long head, whose triangular solve has no bfloat16 CUDA kernel at all.
+        return zc.to(z.dtype), zz[:, -(self.ck - 1):]
 
     def prefill(self, x, state=None):
         if not getattr(self, "ck", 0):
@@ -218,7 +221,7 @@ class ShortLayer(SCA2Layer):
             return super().step(x_t, state)
         z = self.n(x_t)
         win = torch.cat([state["cbuf"].to(z.dtype), z[:, None]], 1)
-        z = (win.transpose(1, 2) * self.cw.squeeze(1).to(z.dtype)).sum(-1)
+        z = (win.transpose(1, 2) * self.cw.squeeze(1).to(z.dtype)).sum(-1).to(win.dtype)
         h = state["z_prev"]
         uc, cs = self.c.step(z, h, state["c"])
         ud, ds = self.dh.step(z, h, state["d"])
